+++ Source/source/game_states/area_editor/gui.cpp
@@ -2141,13 +2141,13 @@ void area_editor::process_gui_panel_sector() {
                         semicolon_list_to_vector(s_ptr->hazards_str);
                     if(
                         selected_hazard_nr >= 0 &&
-                        selected_hazard_nr < list.size()
+                        selected_hazard_nr < (int) list.size()
                     ) {
                         register_change("sector hazard removal");
                         string hazard_name = list[selected_hazard_nr];
                         s_ptr->hazards_str.clear();
                         for(size_t h = 0; h < list.size(); ++h) {
-                            if(h == selected_hazard_nr) continue;
+                            if(h == (size_t) selected_hazard_nr) continue;
                             s_ptr->hazards_str += list[h] + ";";
                         }
                         if(!s_ptr->hazards_str.empty()) {
